Extract var_samp, var_pop pushdown tests to BaseJdbcConnectorTest
